you can now install cli as python package and run it with python3 -m transcendence_cli
