+++ srcs/cli/README.md
@@ -1,0 +1,5 @@
+# Example Package
+
+This is a simple example package. You can use
+[GitHub-flavored Markdown](https://guides.github.com/features/mastering-markdown/)
+to write your content.
