--- srcs/cli/__init__.py
@@ -1,2 +1,0 @@
-from .client import NetworkClient
-from .utils import Singleton, SingletonMeta, UNAUTHORIZED, AUTHORIZED
