+++ srcs/cli/src/transcendence_cli/__init__.py
@@ -1,0 +1,1 @@
+from .cli import main
