+++ srcs/cli/src/transcendence_cli/game.py
@@ -1,4 +1,4 @@
-import client
+from transcendence_cli      import client
 from websockets.sync.client import connect
 import asyncio
 import curses
@@ -25,7 +25,8 @@ class Game:
         self.stdscr.nodelay(1)
         self.key = None
         try:
-            self.ws = connect("ws://c2s15d84.42wolfsburg.de:8000/ws", subprotocols=["Authorization", self.client.access_token])
+            # self.ws = connect("ws://local.42wolfsburg.de:8000/ws", subprotocols=["Authorization", self.client.access_token])
+            self.ws = connect("ws://localhost:8000/ws", additional_headers=[("Authorization", f"Bearer {self.client.access_token}")])
             self.ws.send(json.dumps({"join": self.game_id}))
             # self.waiting_room()
         except Exception as e:
