+++ srcs/cli/src/transcendence_cli/utils.py
@@ -1,5 +1,4 @@
 import curses
-from pickle import TRUE
 UNAUTHORIZED = False
 AUTHORIZED = True
